requirements.txt
# 기존 의존성
opencv-python==4.8.1.78
numpy<2.0
paramiko==3.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
colorama==0.4.6

# 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# 데이터베이스
sqlalchemy==2.0.23
aiosqlite==0.19.0

# 스케줄러
apscheduler==3.10.4

# 알림
aiosmtplib==3.0.1
httpx==0.25.2

# 유틸리티
pydantic==2.5.0
pydantic-settings==2.1.0

